Add setDefaults argument to getValues
Adds the setDefaults argument to getValues to choose whether or not to
store the default values if the keys don't exist. The default value is
set to false, changing the default behaviour of the function. I decided
to do this since it's not likely that you'll need to set the value
before it's actually changed from the default.